fdcb68e96efed9b9210687f2fd02cf810edd21b7 examples/shell.nix: add new commands (Erik Arvstedt)
a2466b11273cadc4fb4a8b5c49937e4b5f2ade3c secrets: allow extending generate-secrets (Erik Arvstedt)
24fd1e9bdceef5c2c21228f2f5a65bb9a2181574 improve examples/shell.nix (Erik Arvstedt)
8a757e0486710f2542e0e3377083d3c595c41436 push-release.sh: improve --dry-run mode (Erik Arvstedt)
82a2b148d8c502d56254b7c6e91731a5cbe41778 secrets: minor fixes (Erik Arvstedt)
e1e3d8a92bf5710098adf0314f4a89005c5d7274 secrets: simplify cert generation (Erik Arvstedt)
2c8e29b35bc5b8165c854966594a68456e341f62 lnd: extract option `certPath` (Erik Arvstedt)
be12a49933aa0034893f4acf06c86766d62f9c87 lightning-pool/loop: extract lnd variable (Erik Arvstedt)
955b44404cc16fefa038f85d4c58929ba558e5be delete helper/fetch-channel (Erik Arvstedt)
5087ce245f15e581e5efb9276806cb1dd9d99d55 minor cleanups (Erik Arvstedt)
0d2db4e79f20324b80e7683a21af7aea89470b06 backups: add option `postgresqlDatabases` (Erik Arvstedt)

Pull request description:

ACKs for top commit:
  nixbitcoin:
    ACK fdcb68e96efed9b9210687f2fd02cf810edd21b7
  jonasnick:
    ACK  fdcb68e96efed9b9210687f2fd02cf810edd21b7

Tree-SHA512: a0fef5b6f8704a445b0e381a1713c14d1447e16798e7035bb005d2b61c4cde208f96fc6f152238b6ea2e9080c04fffe7f841073fa41a5c1e0597204e9ed805c2

+++ helper/push-release.sh
@@ -3,13 +3,9 @@ set -euo pipefail
 
 REPO=fort-nix/nix-bitcoin
 BRANCH=master
-OAUTH_TOKEN=$(pass show nix-bitcoin/github/oauth-token)
+OAUTH_TOKEN=
 DRY_RUN=
 TAG_NAME=
-
-if [[ ! $OAUTH_TOKEN ]]; then
-    echo "Please set OAUTH_TOKEN variable"
-fi
 
 for arg in "$@"; do
     case $arg in
@@ -26,18 +22,28 @@ if [[ ! $TAG_NAME ]]; then
     echo "$0 [--dry-run|-n] <tag_name>"
     exit
 fi
-if [[ $DRY_RUN ]]; then echo "Dry run"; fi
+if [[ $DRY_RUN ]]; then
+    echo "Dry run"
+else
+    OAUTH_TOKEN=$(pass show nix-bitcoin/github/oauth-token)
+    if [[ ! $OAUTH_TOKEN ]]; then
+        echo "Please set OAUTH_TOKEN variable"
+    fi
+fi
 
 RESPONSE=$(curl https://api.github.com/repos/$REPO/releases/latest 2> /dev/null)
 echo "Latest release" $(echo $RESPONSE | jq -r '.tag_name' | tail -c +2)
-while true; do
-    read -p "Create release $TAG_NAME? [yn] " yn
-    case $yn in
-        [Yy]* ) break;;
-        [Nn]* ) exit;;
-        * ) echo "Please answer y or n.";;
-    esac
-done
+
+if [[ ! $DRY_RUN ]]; then
+   while true; do
+       read -p "Create release $TAG_NAME? [yn] " yn
+       case $yn in
+           [Yy]* ) break;;
+           [Nn]* ) exit;;
+           * ) echo "Please answer y or n.";;
+       esac
+   done
+fi
 
 TMPDIR=$(mktemp -d)
 if [[ ! $DRY_RUN ]]; then trap "rm -rf $TMPDIR" EXIT; fi
@@ -52,6 +58,10 @@ SHA256SUMS=$TMPDIR/SHA256SUMS.txt
 # argument
 (cd $TMPDIR; sha256sum $ARCHIVE_NAME > $SHA256SUMS)
 gpg -o $SHA256SUMS.asc -a --detach-sig $SHA256SUMS
+
+cd $TMPDIR
+nix hash to-sri --type sha256 $(nix-prefetch-url --unpack file://$ARCHIVE 2> /dev/null) > nar-hash.txt
+gpg -o nar-hash.txt.asc -a --detach-sig nar-hash.txt
 
 if [[ $DRY_RUN ]]; then
     echo "Created v$TAG_NAME in $TMPDIR"
@@ -71,6 +81,10 @@ post_asset() {
     curl -H "Authorization: token $OAUTH_TOKEN" --data-binary "@$1" -H "Content-Type: application/octet-stream" \
          $GH_ASSET/$(basename $1) &> /dev/null
 }
+post_asset nar-hash.txt
+post_asset nar-hash.txt.asc
+# Post additional assets for backwards compatibility.
+# This allows older nix-bitcoin installations to upgrade via `fetch-release`.
 post_asset $ARCHIVE
 post_asset $SHA256SUMS
 post_asset $SHA256SUMS.asc
